+++ templates/form2.html
@@ -1,0 +1,22 @@
+{% extends "conteneur3.html" %}
+
+{% block titre %}| {{titre}}{%endblock%}
+
+{% block corps %}
+
+   {% include "partials/form.html" %}
+
+    <h1>{{titre}}</h1>
+     <p> Nombre de photo(s) trouvée(s)</p>
+    {% if resultats|length == 1 %}
+    <p>Il y a {{resultats|length}} photo qui correspond à votre requête :</p>
+        <ul>
+            {% for photo in resultats %}
+                        <li><a href="{{url_for('image', id=photo.id)}}">{{photo.titre}}</a></li>
+            {% endfor %}
+        </ul>
+    {% else %}
+    <p>Désolée aucune photo trouvée.</p>
+    {% endif %}
+
+{% endblock %}
