+++ templates/form3.html
@@ -1,0 +1,23 @@
+{% extends "conteneur3.html" %}
+
+{% block titre %}| {{titre}}{%endblock%}
+
+{% block corps %}
+
+   {% include "partials/form.html" %}
+
+    <h1>{{titre}}</h1>
+    {% if resultats|length == 0 %}
+    <p>Désolée aucune photo trouvée.</p>
+    {% elif resultats|length == 1 %}
+    <p>Il y a {{resultats|length}} photo qui correspond à votre requête :</p>
+    {% else %}
+    <p>Il y a {{resultats|length}} photos qui correspondent à votre requête :</p>
+    {% endif %}
+    <ul>
+        {% for photo in resultats %}
+                        <li><a href="{{url_for('image', id=photo.id)}}">{{photo.titre}}</a></li>
+            {% endfor %}
+
+        </ul>
+{% endblock %}
